+++ src/loss.h
@@ -4,8 +4,8 @@
 #include <vector>
 #include "matrix.h"
 
-double cross_entropy(std::vector<float> predictions, int target);
-matrix cross_entropy_derivative(std::vector<float> predictions, int target);
+double cross_entropy(std::vector<double> predictions, int target);
+matrix cross_entropy_derivative(std::vector<double> predictions, int target);
 
 
 #endif //NEURAL_NETWORK_LOSS_H
